Avoid too many values to unpack error
I just installed this plugin to see what it can do with my EyeTV Hybrid before I decide what hardware/software to buy for OTA TV. The EyeTV-parser add-on would not launch and the log showed a trace for an error coming from this line (38) saying there were too many values to unpack, so I printed the args to debug and it contained 4 values, the last being 'resume:false'. Supplying the dummy value allowed the plugin to load.  There are of course other issues, but this was a nice/simple un-blocking work-around/fix.